Added addRestNameChain.c and changed code[0][X-1] and code[X-1][X-1] to be rings.
To do this, processes would be faster.
Because, for example, the initial value of code[0][X-1] can be set as code[X-1][0]+1.

So far, there's a bug that makes X=4 result '3520'.
This is caused by the change on code[0][X-1] and code[X-1][X-1] to be rings, not afills.
However, the relationship between code[0][0] and code[X-1][0] is eternal
So the result value is 880*'4' instead of 880*'8' in X=4.

I will not fix this bug.
Instead, I will make the code which set the corner code value in smart way.